+++ app/src/main/java/ru/practicum/android/diploma/core/data/db/converter/VacancyTypeConverters.kt
@@ -1,15 +1,14 @@
 package ru.practicum.android.diploma.core.data.db.converter
 
+import androidx.room.ProvidedTypeConverter
 import androidx.room.TypeConverter
 import com.google.gson.Gson
-import com.google.gson.GsonBuilder
 import ru.practicum.android.diploma.core.domain.models.VacancyDetails
 
-class VacancyTypeConverters {
-
-    private val gson: Gson = GsonBuilder()
-        .serializeNulls()
-        .create()
+@ProvidedTypeConverter
+class VacancyTypeConverters(
+    private val gson: Gson
+) {
 
     @TypeConverter
     fun fromVacancyDetails(vacancyDetails: VacancyDetails): String {
